requirements.txt
# CSC111 Winter 2023: Python libraries we'll be using this semester.

# Testing and code checking
hypothesis==6.68.2
pytest==7.2.1
python-ta==2.4.2

# Graphics and data visualization
plotly==5.13.1
pygame~=2.1.3.dev8
networkx~=3.0
matplotlib~=3.7.1

Pillow~=9.4.0
